# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenThreatGrid API",
    "summary": "Ingestion and query API for honeypot threat telemetry.",
    "description": "Ingestion and query API for **OpenThreatGrid** honeypot threat telemetry.\n\nSensors (Cowrie first) are parsed and enriched by the worker, then submitted\nhere as normalized **OTG Standard Events**. Events are validated, stored in\nOpenSearch (`otg-events-*`), and exposed for dashboards and the weekly threat\nreport.\n\n### Pipeline\n\n```\nCowrie \u2192 parser \u2192 Redis \u2192 consumer (enrich + GeoIP/ASN) \u2192 THIS API \u2192 OpenSearch\n```\n\n### Conventions\n\n* **Ingestion is idempotent** \u2014 submit the same `event_id` twice and the\n  duplicate is skipped (reported as not accepted), so a worker can safely retry\n  a batch.\n* Timestamps are ISO-8601 (UTC). See the\n  [event schema](https://github.com/OpenThreatGrid/openthreatgrid/blob/main/docs/event-schema.md).\n* Stats endpoints are backed by OpenSearch aggregations and power the\n  dashboards and reports.\n",
    "contact": {
      "name": "OpenThreatGrid",
      "url": "https://github.com/OpenThreatGrid/openthreatgrid"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.openthreatgrid.io",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Ingest Events",
        "description": "Ingest a single event or a batch of events.\n\nAccepts either a JSON object or a JSON array. Duplicate ``event_id`` values\nare skipped so a worker can safely retry a batch without creating\nduplicates.",
        "operationId": "ingest_events_api_v1_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EventCreate"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/EventCreate"
                    }
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "description": "Query events with optional filters, newest first.",
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "source_ip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Ip"
            }
          },
          {
            "name": "sensor_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventRead"
                  },
                  "title": "Response List Events Api V1 Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "description": "Fetch a single event by its ``event_id``.",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/summary": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Stats Summary",
        "description": "Aggregated counts across all stored events.",
        "operationId": "stats_summary_api_v1_stats_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/top-source-ips": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Top Source Ips",
        "operationId": "top_source_ips_api_v1_stats_top_source_ips_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Count"
                  },
                  "title": "Response Top Source Ips Api V1 Stats Top Source Ips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/top-usernames": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Top Usernames",
        "operationId": "top_usernames_api_v1_stats_top_usernames_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Count"
                  },
                  "title": "Response Top Usernames Api V1 Stats Top Usernames Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/top-passwords": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Top Passwords",
        "operationId": "top_passwords_api_v1_stats_top_passwords_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Count"
                  },
                  "title": "Response Top Passwords Api V1 Stats Top Passwords Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/top-commands": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Top Commands",
        "operationId": "top_commands_api_v1_stats_top_commands_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Count"
                  },
                  "title": "Response Top Commands Api V1 Stats Top Commands Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health",
        "description": "Liveness probe \u2014 process is up.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Ready",
        "description": "Readiness probe \u2014 verifies OpenSearch (and Redis if configured) connectivity.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Ready Ready Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Count": {
        "properties": {
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "key",
          "count"
        ],
        "title": "Count",
        "description": "A single labelled count, used for top-N breakdowns."
      },
      "EventBatchResult": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Ids"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "event_ids"
        ],
        "title": "EventBatchResult",
        "description": "Response for a (possibly batched) ingestion request.\n\n``accepted`` is the number of *newly stored* events; duplicates (by\n``event_id``) are silently skipped, so ``accepted`` may be less than the\nnumber submitted.",
        "example": {
          "accepted": 1,
          "event_ids": [
            "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          ]
        }
      },
      "EventCreate": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "sensor_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Sensor Id"
          },
          "sensor_type": {
            "type": "string",
            "maxLength": 64,
            "title": "Sensor Type",
            "default": "cowrie"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "source_ip": {
            "type": "string",
            "maxLength": 45,
            "title": "Source Ip"
          },
          "source_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Port"
          },
          "destination_ip": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 45
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Ip"
          },
          "destination_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Port"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "payload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Url"
          },
          "payload_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Hash"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "geo_country": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Country"
          },
          "geo_asn": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Asn"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "raw_event": {
            "type": "object",
            "title": "Raw Event"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "sensor_id",
          "event_type",
          "source_ip"
        ],
        "title": "EventCreate",
        "description": "Schema for ingesting an event.\n\n``event_id`` is optional on input \u2014 if the producer (worker) does not set\none, the API generates a UUIDv4 so events are idempotently identifiable.",
        "example": {
          "destination_port": 22,
          "event_type": "login_attempt",
          "password": "admin123",
          "protocol": "ssh",
          "sensor_id": "cowrie-sg-01",
          "sensor_type": "cowrie",
          "source_ip": "203.0.113.42",
          "source_port": 54321,
          "success": false,
          "tags": [
            "ssh",
            "bruteforce"
          ],
          "timestamp": "2026-06-01T12:34:56.789Z",
          "username": "root"
        }
      },
      "EventRead": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "sensor_id": {
            "type": "string",
            "maxLength": 128,
            "title": "Sensor Id"
          },
          "sensor_type": {
            "type": "string",
            "maxLength": 64,
            "title": "Sensor Type",
            "default": "cowrie"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "source_ip": {
            "type": "string",
            "maxLength": 45,
            "title": "Source Ip"
          },
          "source_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Port"
          },
          "destination_ip": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 45
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Ip"
          },
          "destination_port": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 65535.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Port"
          },
          "protocol": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "payload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Url"
          },
          "payload_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Hash"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "geo_country": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Country"
          },
          "geo_asn": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Asn"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "raw_event": {
            "type": "object",
            "title": "Raw Event"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "ingested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingested At"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "sensor_id",
          "event_type",
          "source_ip",
          "event_id"
        ],
        "title": "EventRead",
        "description": "Schema returned when querying events."
      },
      "EventType": {
        "type": "string",
        "enum": [
          "login_attempt",
          "login_success",
          "command_exec",
          "file_download",
          "session_start",
          "session_close",
          "connection"
        ],
        "title": "EventType",
        "description": "Supported OpenThreatGrid event types (see docs/event-schema.md)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "StatsSummary": {
        "properties": {
          "total_events": {
            "type": "integer",
            "title": "Total Events"
          },
          "unique_source_ips": {
            "type": "integer",
            "title": "Unique Source Ips"
          },
          "unique_sensors": {
            "type": "integer",
            "title": "Unique Sensors"
          },
          "events_by_type": {
            "items": {
              "$ref": "#/components/schemas/Count"
            },
            "type": "array",
            "title": "Events By Type"
          },
          "top_source_ips": {
            "items": {
              "$ref": "#/components/schemas/Count"
            },
            "type": "array",
            "title": "Top Source Ips"
          },
          "top_usernames": {
            "items": {
              "$ref": "#/components/schemas/Count"
            },
            "type": "array",
            "title": "Top Usernames"
          },
          "top_passwords": {
            "items": {
              "$ref": "#/components/schemas/Count"
            },
            "type": "array",
            "title": "Top Passwords"
          },
          "first_event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Event"
          },
          "last_event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event"
          }
        },
        "type": "object",
        "required": [
          "total_events",
          "unique_source_ips",
          "unique_sensors",
          "events_by_type",
          "top_source_ips",
          "top_usernames",
          "top_passwords"
        ],
        "title": "StatsSummary",
        "description": "Aggregated view of stored events, powering the dashboard and reports."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "events",
      "description": "Ingest and query normalized honeypot events."
    },
    {
      "name": "stats",
      "description": "Aggregated counts and top-N breakdowns for dashboards and reports."
    },
    {
      "name": "meta",
      "description": "Liveness and readiness probes."
    }
  ]
}